Remove ability to edit profile. Redirect to OneSage
