+++ src/main/resources/org/sagebionetworks/web/client/widget/profile/UserProfileWidgetViewImpl.ui.xml
@@ -29,15 +29,13 @@
           <b:Column size="MD_12,LG_6">
             <b:FormGroup>
               <b:FormLabel for="firstName">First Name</b:FormLabel>
-              <b:TextBox name="firstName" ui:field="firstName" />
-              <bh:Paragraph ui:field="firstNameRenderer" visible="false" />
+              <bh:Paragraph ui:field="firstNameRenderer" />
             </b:FormGroup>
           </b:Column>
           <b:Column size="MD_12,LG_6">
             <b:FormGroup>
               <b:FormLabel for="lastName">Last Name</b:FormLabel>
-              <b:TextBox name="lastName" ui:field="lastName" />
-              <bh:Paragraph ui:field="lastNameRenderer" visible="false" />
+              <bh:Paragraph ui:field="lastNameRenderer" />
             </b:FormGroup>
           </b:Column>
         </b:Row>
@@ -45,11 +43,7 @@
           <b:Column size="MD_12,LG_6">
             <b:FormGroup>
               <b:FormLabel for="currentPosition">Current Position</b:FormLabel>
-              <b:TextBox name="currentPosition" ui:field="currentPosition" />
-              <bh:Paragraph
-                ui:field="currentPositionRenderer"
-                visible="false"
-              />
+              <bh:Paragraph ui:field="currentPositionRenderer" />
             </b:FormGroup>
           </b:Column>
           <b:Column size="MD_12,LG_6">
@@ -57,14 +51,7 @@
               <b:FormLabel for="currentAffiliation">
                 Current Affiliation
               </b:FormLabel>
-              <b:TextBox
-                name="currentAffiliation"
-                ui:field="currentAffiliation"
-              />
-              <bh:Paragraph
-                ui:field="currentAffiliationRenderer"
-                visible="false"
-              />
+              <bh:Paragraph ui:field="currentAffiliationRenderer" />
             </b:FormGroup>
           </b:Column>
         </b:Row>
@@ -72,11 +59,9 @@
           <b:Column size="XS_12">
             <b:FormGroup>
               <b:FormLabel for="bio">Bio</b:FormLabel>
-              <b:TextArea name="bio" ui:field="bioEditor" height="150px" />
               <bh:Paragraph
                 ui:field="bioRenderer"
                 addStyleNames="whitespace-prewrap"
-                visible="false"
               />
             </b:FormGroup>
           </b:Column>
@@ -85,8 +70,7 @@
           <b:Column size="MD_12,LG_6">
             <b:FormGroup ui:field="usernameFormGroup">
               <b:FormLabel for="username">Username</b:FormLabel>
-              <b:TextBox name="username" ui:field="username" />
-              <bh:Paragraph ui:field="usernameRenderer" visible="false" />
+              <bh:Paragraph ui:field="usernameRenderer" />
               <b:HelpBlock ui:field="usernameHelpBlock" />
             </b:FormGroup>
           </b:Column>
@@ -96,7 +80,6 @@
               <bh:Div>
                 <b:Anchor ui:field="linkRenderer" target="_blank" />
               </bh:Div>
-              <b:TextBox name="link" ui:field="link" />
               <b:HelpBlock ui:field="linkHelpBlock" />
             </b:FormGroup>
           </b:Column>
@@ -105,15 +88,13 @@
           <b:Column size="MD_12,LG_6">
             <b:FormGroup>
               <b:FormLabel for="industry">Industry/Discipline</b:FormLabel>
-              <b:TextBox name="industry" ui:field="industry" />
-              <bh:Paragraph ui:field="industryRenderer" visible="false" />
+              <bh:Paragraph ui:field="industryRenderer" />
             </b:FormGroup>
           </b:Column>
           <b:Column size="MD_12,LG_6">
             <b:FormGroup>
               <b:FormLabel for="location">City, Country</b:FormLabel>
-              <bh:Div ui:field="locationSuggestBoxContainer" />
-              <bh:Paragraph ui:field="locationRenderer" visible="false" />
+              <bh:Paragraph ui:field="locationRenderer" />
             </b:FormGroup>
           </b:Column>
         </b:Row>
@@ -126,15 +107,6 @@
                 marginBottom="10"
                 addStyleNames="truncate"
               />
-              <b:Anchor
-                addStyleNames="displayInlineBlock"
-                ui:field="changeEmailLink"
-                visible="false"
-                target="_blank"
-                href="https://accounts.sagebionetworks.synapse.org/authenticated/myaccount"
-              >
-                Change email
-              </b:Anchor>
             </b:FormGroup>
           </b:Column>
           <b:Column size="MD_12,LG_6" ui:field="orcIDContainer">
@@ -143,30 +115,6 @@
               <bh:Div>
                 <b:Anchor ui:field="orcIdLink" target="_blank" />
               </bh:Div>
-            </b:FormGroup>
-          </b:Column>
-        </b:Row>
-        <b:Row ui:field="passwordContainer">
-          <b:Column size="MD_12,LG_6">
-            <b:FormGroup>
-              <b:FormLabel for="password">Password</b:FormLabel>
-              <g:PasswordTextBox
-                name="password"
-                readOnly="true"
-                addStyleNames="displayBlock margin-bottom-10"
-              >
-                NotThePassword
-              </g:PasswordTextBox>
-
-              <b:Anchor
-                addStyleNames="displayInlineBlock"
-                ui:field="changePasswordLink"
-                target="_blank"
-                href="https://accounts.sagebionetworks.synapse.org/authenticated/myaccount#ChangePassword"
-                visible="false"
-              >
-                Change password
-              </b:Anchor>
             </b:FormGroup>
           </b:Column>
         </b:Row>
@@ -184,21 +132,18 @@
 
         <bh:Div ui:field="synAlertContainer" />
         <bh:Div marginTop="45" marginBottom="50" ui:field="commandsContainer">
-          <b:Button ui:field="cancelButton" type="DEFAULT" visible="false">
-            Cancel
-          </b:Button>
-          <b:Button
-            ui:field="saveProfileButton"
-            type="PRIMARY"
-            visible="false"
-            dataLoadingText="Saving Changes..."
-            marginLeft="5"
-          >
-            Save Changes
-          </b:Button>
-          <b:Button ui:field="editProfileButton" type="PRIMARY" visible="false">
+          <b:Button ui:field="editProfileButton" type="PRIMARY">
             Edit Profile
           </b:Button>
+          <b:Anchor
+            addStyleNames="displayBlock"
+            marginTop="10"
+            ui:field="changePasswordLink"
+            target="_blank"
+            href="https://accounts.sagebionetworks.synapse.org/authenticated/myaccount#ChangePassword"
+          >
+            Change password
+          </b:Anchor>
         </bh:Div>
       </bh:Div>
       <bh:Div
