+++ src/test/java/org/sagebionetworks/web/unitclient/widget/profile/UserProfileWidgetImplTest.java
@@ -1,11 +1,7 @@
 package org.sagebionetworks.web.unitclient.widget.profile;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertTrue;
 import static org.mockito.Matchers.*;
 import static org.mockito.Mockito.never;
-import static org.mockito.Mockito.reset;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
@@ -35,17 +31,16 @@ import org.sagebionetworks.web.client.security.AuthenticationController;
 import org.sagebionetworks.web.client.utils.Callback;
 import org.sagebionetworks.web.client.widget.entity.controller.SynapseAlert;
 import org.sagebionetworks.web.client.widget.profile.ProfileImageWidget;
-import org.sagebionetworks.web.client.widget.profile.UserProfileEditorWidgetImpl;
-import org.sagebionetworks.web.client.widget.profile.UserProfileEditorWidgetView;
+import org.sagebionetworks.web.client.widget.profile.UserProfileWidgetImpl;
+import org.sagebionetworks.web.client.widget.profile.UserProfileWidgetView;
 import org.sagebionetworks.web.client.widget.upload.CroppedImageUploadViewImpl;
 import org.sagebionetworks.web.client.widget.upload.ImageUploadWidget;
-import org.sagebionetworks.web.shared.WebConstants;
 import org.sagebionetworks.web.test.helper.AsyncMockStubber;
 
-public class UserProfileEditorWidgetImplTest {
+public class UserProfileWidgetImplTest {
 
   @Mock
-  UserProfileEditorWidgetView mockView;
+  UserProfileWidgetView mockView;
 
   @Mock
   ProfileImageWidget mockImageWidget;
@@ -53,7 +48,7 @@ public class UserProfileEditorWidgetImplTest {
   @Mock
   ImageUploadWidget mockfileHandleUploadWidget;
 
-  UserProfileEditorWidgetImpl widget;
+  UserProfileWidgetImpl widget;
 
   @Mock
   PortalGinInjector mockPortalGinInjector;
@@ -112,7 +107,7 @@ public class UserProfileEditorWidgetImplTest {
     when(mockAuthController.getCurrentUserPrincipalId())
       .thenReturn(USER_PROFILE_ID);
     widget =
-      new UserProfileEditorWidgetImpl(
+      new UserProfileWidgetImpl(
         mockView,
         mockImageWidget,
         mockfileHandleUploadWidget,
@@ -159,17 +154,6 @@ public class UserProfileEditorWidgetImplTest {
     changes.setProfilePicureFileHandleId("456782");
     changes.setDisplayName("James-2 Bond-2");
 
-    widget.setNewFileHandle(changes.getProfilePicureFileHandleId());
-    when(mockView.getUsername()).thenReturn(changes.getUserName());
-    when(mockView.getFirstName()).thenReturn(changes.getFirstName());
-    when(mockView.getLastName()).thenReturn(changes.getLastName());
-    when(mockView.getCurrentPosition()).thenReturn(changes.getPosition());
-    when(mockView.getCurrentAffiliation()).thenReturn(changes.getCompany());
-    when(mockView.getIndustry()).thenReturn(changes.getIndustry());
-    when(mockView.getLocation()).thenReturn(changes.getLocation());
-    when(mockView.getLink()).thenReturn(changes.getUrl());
-    when(mockView.getBio()).thenReturn(changes.getSummary());
-
     when(mockGlobalAppState.getPlaceChanger()).thenReturn(mockPlaceChanger);
     AsyncMockStubber
       .callSuccessWith(profile)
@@ -194,15 +178,8 @@ public class UserProfileEditorWidgetImplTest {
     verify(mockView).setSynAlert(mockSynAlert);
   }
 
-  private void verifyIsEditingMode(boolean isEditing, int nCalls) {
-    verify(mockGlobalAppState, times(nCalls)).setIsEditing(isEditing);
-    verify(mockView, times(nCalls)).setEditMode(isEditing);
-  }
-
   private void verifyConfigure(int nCalls) {
     verify(mockSynAlert, times(nCalls)).clear();
-    verify(mockView, times(nCalls)).hideLinkError();
-    verify(mockView, times(nCalls)).hideUsernameError();
     verify(mockView, times(nCalls)).clearEmails();
     verify(mockView, times(nCalls)).setOwnerId(profile.getOwnerId());
     verify(mockView, times(nCalls)).setFirstName(profile.getFirstName());
@@ -217,7 +194,6 @@ public class UserProfileEditorWidgetImplTest {
     verify(mockView, times(nCalls)).setBio(profile.getSummary());
     verify(mockImageWidget, times(nCalls))
       .configure(profile.getOwnerId(), profile.getProfilePicureFileHandleId());
-    verifyIsEditingMode(false, nCalls);
   }
 
   @Test
@@ -264,129 +240,9 @@ public class UserProfileEditorWidgetImplTest {
   }
 
   @Test
-  public void testOnCancel() {
-    widget.configure(profile, ORC_ID, mockCallback);
-    widget.onCancel();
-
-    verifyConfigure(2);
-  }
-
-  @Test
-  public void testOnSave() {
-    widget.configure(profile, null, mockCallback);
-    widget.onSave();
-
-    verify(mockSynAlert, times(2)).clear();
-    verify(mockJsClient)
-      .updateMyUserProfile(
-        userProfileCaptor.capture(),
-        any(AsyncCallback.class)
-      );
-
-    //verify profile was updated to new values
-    assertEquals(
-      changes.getOwnerId(),
-      userProfileCaptor.getValue().getOwnerId()
-    );
-    assertEquals(
-      changes.getFirstName(),
-      userProfileCaptor.getValue().getFirstName()
-    );
-    assertEquals(
-      changes.getLastName(),
-      userProfileCaptor.getValue().getLastName()
-    );
-    assertEquals(
-      changes.getUserName(),
-      userProfileCaptor.getValue().getUserName()
-    );
-
-    verify(mockCallback).invoke();
-    verify(mockClientCache)
-      .remove(profile.getOwnerId() + WebConstants.USER_PROFILE_SUFFIX);
-    verify(mockAuthController).updateCachedProfile(profile);
-    verify(mockPlaceChanger, never()).goTo(any(Profile.class));
-  }
-
-  @Test
-  public void testOnSaveFailure() {
-    Exception ex = new Exception("An error");
-    AsyncMockStubber
-      .callFailureWith(ex)
-      .when(mockJsClient)
-      .updateMyUserProfile(any(UserProfile.class), any(AsyncCallback.class));
-
-    widget.configure(profile, ORC_ID, mockCallback);
-    widget.onSave();
-
-    verify(mockSynAlert, times(2)).clear();
-    verify(mockJsClient)
-      .updateMyUserProfile(
-        userProfileCaptor.capture(),
-        any(AsyncCallback.class)
-      );
-    verify(mockSynAlert).handleException(ex);
-    verify(mockView).resetSaveButtonState();
-  }
-
-  @Test
-  public void testSetNewFileHandle() {
-    String newFileHandle = "293";
-    widget.setUploadingCompleteCallback(mockCallback);
-    verify(mockCallback, never()).invoke();
-
-    widget.setNewFileHandle(newFileHandle);
-
-    verify(mockImageWidget).configure(newFileHandle);
-    verify(mockCallback).invoke();
-  }
-
-  @Test
-  public void testSetUploadingCallback() {
-    widget.setUploadingCallback(mockCallback);
-    verify(mockfileHandleUploadWidget).setUploadingCallback(mockCallback);
-  }
-
-  @Test
   public void testConfigureNoProfileImage() {
     profile.setProfilePicureFileHandleId(null);
     widget.configure(profile, ORC_ID, mockCallback);
     verify(mockImageWidget).configure(profile.getOwnerId(), null);
-  }
-
-  @Test
-  public void testIsValid() {
-    widget.configure(profile, ORC_ID, mockCallback);
-    reset(mockView);
-    when(mockView.getUsername()).thenReturn("valid");
-    assertTrue(widget.isValid());
-    verify(mockView).hideLinkError();
-    verify(mockView).hideUsernameError();
-  }
-
-  @Test
-  public void testIsValidShortUsername() {
-    widget.configure(profile, ORC_ID, mockCallback);
-    reset(mockView);
-    when(mockView.getUsername()).thenReturn("12");
-    assertFalse(widget.isValid());
-    verify(mockView).hideLinkError();
-    verify(mockView).hideUsernameError();
-    verify(mockView)
-      .showUsernameError(
-        UserProfileEditorWidgetImpl.MUST_BE_AT_LEAST_3_CHARACTERS
-      );
-  }
-
-  @Test
-  public void testIsValidUsernameBadChars() {
-    widget.configure(profile, ORC_ID, mockCallback);
-    reset(mockView);
-    when(mockView.getUsername()).thenReturn("ABC@");
-    assertFalse(widget.isValid());
-    verify(mockView).hideLinkError();
-    verify(mockView).hideUsernameError();
-    verify(mockView)
-      .showUsernameError(UserProfileEditorWidgetImpl.CAN_ONLY_INCLUDE);
   }
 }
